Revert "Started Services and Actions for #1689 (#1871)"
This reverts commit ed748e2779fe0d8fd74726c5a0a5e7f0de93b27e.
