+++ app/controllers/usage_controller.rb
@@ -2,8 +2,12 @@
 
 class UsageController < ApplicationController
 
+  # GET /usage
   def index
-    check_authorized!
+    unless current_user.present? &&
+          (current_user.can_org_admin? || current_user.can_super_admin?)
+      raise Pundit::NotAuthorizedError
+    end
     render("index",
       locals: {
         orgs: Org.all,
@@ -13,12 +17,4 @@ class UsageController < ApplicationController
     )
   end
 
-  private
-
-  def check_authorized!
-    unless current_user.present? &&
-        (current_user.can_org_admin? || current_user.can_super_admin?)
-      raise Pundit::NotAuthorizedError
-    end
-  end
 end
