+++ config/routes.rb
@@ -125,8 +125,6 @@ Rails.application.routes.draw do
 
   resources :usage, only: [:index]
 
-  resources :usage_downloads, only: [:index]
-
   resources :roles, only: [:create, :update, :destroy] do
     member do
       put :deactivate
