+++ spec/models/stat_created_plan_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe StatCreatedPlan, type: :model do
         csv = described_class.to_csv(data)
 
         expected_csv = <<~HERE
-          Date,Count
+          date,count
           2018-05-31,20
           2018-06-30,10
         HERE
